add process to scheduler & update priority accordingly
